feat(Routes): routes restructure; integrated route state in dho page tabs;
