+++ apps/web/src/app/[lang]/my-spaces/[id]/assignments/@modal/default.tsx
