+++ apps/web/src/app/[lang]/my-spaces/[id]/assignments/constants.ts
@@ -1,6 +1,6 @@
 import { Locale } from '@hypha-platform/i18n';
 import { PATH_DHO, PATH_SEGMENT_DHO_ID } from '../constants';
-import { PATH_SEGMENT_LANG } from '../../../constants';
+import { PATH_SEGMENT_LANG } from '../../../../constants';
 
 export const PATH_ASSIGNMENTS = `${PATH_DHO}/assignments`;
