+++ apps/web/src/app/[lang]/my-spaces/[id]/assignments/create/constants.ts
