+++ apps/web/src/app/[lang]/my-spaces/[id]/assignments/layout.tsx
