+++ apps/web/src/app/[lang]/my-spaces/[id]/constants.ts
@@ -4,6 +4,6 @@ export const PATH_SEGMENT_DHO_ID = ':id';
 
 export const PATH_DHO = `${PATH_LANG}/dho/${PATH_SEGMENT_DHO_ID}`;
 
-export const getDhoPath = (lang: Locale, id: string) => {
-  return `/${lang}/dho/${id}`;
+export const getDhoPathAgreements = (lang: Locale, id: string) => {
+  return `/${lang}/my-spaces/${id}/agreements`;
 };
