+++ apps/web/src/app/[lang]/my-spaces/new/layout.tsx
